bush script

+++ source/tester.cpp
@@ -13,7 +13,7 @@ void Tester(){
                             {7, 1, 1,  0, -1, 0,  2},
                             {8, 1, 1, -2, -2, 1,  2} };
 
-    for (int i = 0; i < (sizeof(list) / sizeof(list[0])); ++i) {
+    for (unsigned long i = 0; i < (sizeof(list) / sizeof(list[0])); ++i) {
         printf("%s\n", TestsMessenger(RunTests(&list[i])));
     }
 }
